Clean up a bunch of warnings

+++ utility.cpp
@@ -348,7 +348,7 @@ json::JSON json::JSON::Load(const string& str) {
 
 json::JSON EncodeDWORDArray(DWORD* dwArray, size_t iCount, BOOL bBigEndian) {
 	json::JSON jsonArray = json::Array();
-	for (int i = 0; i < iCount; i++) {
+	for (size_t i = 0; i < iCount; i++) {
 		if (bBigEndian)
 			jsonArray.append<DWORD>(SwapDWORD(dwArray[i]));
 		else
@@ -359,6 +359,6 @@ json::JSON EncodeDWORDArray(DWORD* dwArray, size_t iCount, BOOL bBigEndian) {
 
 // Scary function! Overflows abound! Be careful!
 void DecodeDWORDArray(DWORD* dwArray, json::JSON jsonArray, size_t iCount, BOOL bBigEndian) {
-	for (int i = 0; i < iCount; i++)
+	for (size_t i = 0; i < iCount; i++)
 		dwArray[i] = (bBigEndian ? SwapDWORD(jsonArray[i].ToInt()) : jsonArray[i].ToInt());
 }
